agents/Competitor analysis agent/requirements.txt
crewai==0.28.8
crewai-tools==0.1.6
langchain==0.1.20
langchain-openai==0.1.7
google-search-results==2.4.2
tweepy==4.14.0
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
slack-sdk==3.27.1
beautifulsoup4==4.12.3
lxml==5.1.0
